add bcm27xx (RaspberryPi) target
The supported devices are grouped into 3 profiles:
- default:
  + RPi 0/0W
  + RPi A/A+/B/B+
- bcm2709:
  + RPi 2 (v1.0, v1.1, v1.2)
- bcm2710:
  + RPi 3A/3A+/3B/3B+

The default and bcm2709 squashfs images have been tested on an RPi B+
and RPi 2 v1.2 respectively.  bcm2709 squashfs sysupgrade has also
been tested on an RPi 2 v1.2.  In principle the bcm2709 images should
also work for RPI 3 boards as well but this hasn't been tested.  The
ext4 images have not been tested.

I would recommend using the squashfs images with SD cards to avoid
problems with restarting after power loss (when ext4 filesystems need
to be fsck'ed) and to minimise write wear on the flash components.
The ext4 images would be better suited to setups with USB drives.

All profiles have been configured functionally alike with most Gargoyle
functional plugins included.  As these images are not particularly
size constrained, all languages and all themes have been included.
To support wired router operation, all images include support for
common USB ethernet interfaces (Asix AXE and Realtek r8150/r8152
- AXE 88179 and r8152 tested) as well as USB 4G/LTE modems (Huawei
E3372h tested).

On devices with a built-in ethernet port, that port will be used
as the LAN interface.  If a USB ethernet interface is plugged
in and detected on first boot on devices with a built-in ethernet
port, the USB ethernet interface will be used as the WAN interface.