package/plugin-gargoyle-email-notifications: align SSL dependencies
The PROVIDES=msmtp for the nossl variant of the msmtp package defeats
the dependency for the SSL variant of the plugin, resulting in only
the nossl variant of the msmtp package being built by default.

Fix this by deactivating the PROVIDES and adding explicit msmtp package
variant selection in both plugin variants to match the SSL variant.
NB: PROVIDES has been commented out rather than removed to ease
    restoration in any future restructure of these packages.

Signed-off-by: Andrew MacIntyre <andymac@pcug.org.au>